inference_mnist/requirements.txt
numpy==1.26.4
requests==2.32.3
pillow==11.1.0
tritonclient[grpc]==2.53.0